chore: fix some examples
