--- examples/multi_user_chat/group_hierarchy/start.yml
@@ -1,4 +1,0 @@
-agents:
-  agent:
-    file: agent.py
-    domain: localhost
